Updated JWT + comments
